ops: restructure project files and create helper
