--- hc/.vscode/settings.json
@@ -1,6 +1,0 @@
-{
-  "[python]": {
-    "editor.defaultFormatter": "ms-python.black-formatter"
-  },
-  "python.formatting.provider": "none"
-}
