+++ input/st70.tsp.txt
