--- sa/input/bier127.tsp.txt
@@ -1,127 +1,0 @@
-  1   9860  14152
-   2   9396  14616
-   3  11252  14848
-   4  11020  13456
-   5   9512  15776
-   6  10788  13804
-   7  10208  14384
-   8  11600  13456
-   9  11252  14036
-  10  10672  15080
-  11  11136  14152
-  12   9860  13108
-  13  10092  14964
-  14   9512  13340
-  15  10556  13688
-  16   9628  14036
-  17  10904  13108
-  18  11368  12644
-  19  11252  13340
-  20  10672  13340
-  21  11020  13108
-  22  11020  13340
-  23  11136  13572
-  24  11020  13688
-  25   8468  11136
-  26   8932  12064
-  27   9512  12412
-  28   7772  11020
-  29   8352  10672
-  30   9164  12876
-  31   9744  12528
-  32   8352  10324
-  33   8236  11020
-  34   8468  12876
-  35   8700  14036
-  36   8932  13688
-  37   9048  13804
-  38   8468  12296
-  39   8352  12644
-  40   8236  13572
-  41   9164  13340
-  42   8004  12760
-  43   8584  13108
-  44   7772  14732
-  45   7540  15080
-  46   7424  17516
-  47   8352  17052
-  48   7540  16820
-  49   7888  17168
-  50   9744  15196
-  51   9164  14964
-  52   9744  16240
-  53   7888  16936
-  54   8236  15428
-  55   9512  17400
-  56   9164  16008
-  57   8700  15312
-  58  11716  16008
-  59  12992  14964
-  60  12412  14964
-  61  12296  15312
-  62  12528  15196
-  63  15312   6612
-  64  11716  16124
-  65  11600  19720
-  66  10324  17516
-  67  12412  13340
-  68  12876  12180
-  69  13688  10904
-  70  13688  11716
-  71  13688  12528
-  72  11484  13224
-  73  12296  12760
-  74  12064  12528
-  75  12644  10556
-  76  11832  11252
-  77  11368  12296
-  78  11136  11020
-  79  10556  11948
-  80  10324  11716
-  81  11484   9512
-  82  11484   7540
-  83  11020   7424
-  84  11484   9744
-  85  16936  12180
-  86  17052  12064
-  87  16936  11832
-  88  17052  11600
-  89  13804  18792
-  90  12064  14964
-  91  12180  15544
-  92  14152  18908
-  93   5104  14616
-  94   6496  17168
-  95   5684  13224
-  96  15660  10788
-  97   5336  10324
-  98    812   6264
-  99  14384  20184
- 100  11252  15776
- 101   9744   3132
- 102  10904   3480
- 103   7308  14848
- 104  16472  16472
- 105  10440  14036
- 106  10672  13804
- 107   1160  18560
- 108  10788  13572
- 109  15660  11368
- 110  15544  12760
- 111   5336  18908
- 112   6264  19140
- 113  11832  17516
- 114  10672  14152
- 115  10208  15196
- 116  12180  14848
- 117  11020  10208
- 118   7656  17052
- 119  16240   8352
- 120  10440  14732
- 121   9164  15544
- 122   8004  11020
- 123   5684  11948
- 124   9512  16472
- 125  13688  17516
- 126  11484   8468
- 127   3248  14152
