--- sa/input/kroE100.tsp.txt
@@ -1,100 +1,0 @@
-1 3477 949
-2 91 1732
-3 3972 329
-4 198 1632
-5 1806 733
-6 538 1023
-7 3430 1088
-8 2186 766
-9 1513 1646
-10 2143 1611
-11 53 1657
-12 3404 1307
-13 1034 1344
-14 2823 376
-15 3104 1931
-16 3232 324
-17 2790 1457
-18 374 9
-19 741 146
-20 3083 1938
-21 3502 1067
-22 1280 237
-23 3326 1846
-24 217 38
-25 2503 1172
-26 3527 41
-27 739 1850
-28 3548 1999
-29 48 154
-30 1419 872
-31 1689 1223
-32 3468 1404
-33 1628 253
-34 382 872
-35 3029 1242
-36 3646 1758
-37 285 1029
-38 1782 93
-39 1067 371
-40 2849 1214
-41 920 1835
-42 1741 712
-43 876 220
-44 2753 283
-45 2609 1286
-46 3941 258
-47 3613 523
-48 1754 559
-49 2916 1724
-50 2445 1820
-51 3825 1101
-52 2779 435
-53 201 693
-54 2502 1274
-55 765 833
-56 3105 1823
-57 1937 1400
-58 3364 1498
-59 3702 1624
-60 2164 1874
-61 3019 189
-62 3098 1594
-63 3239 1376
-64 3359 1693
-65 2081 1011
-66 1398 1100
-67 618 1953
-68 1878 59
-69 3803 886
-70 397 1217
-71 3035 152
-72 2502 146
-73 3230 380
-74 3479 1023
-75 958 1670
-76 3423 1241
-77 78 1066
-78 96 691
-79 3431 78
-80 2053 1461
-81 3048 1
-82 571 1711
-83 3393 782
-84 2835 1472
-85 144 1185
-86 923 108
-87 989 1997
-88 3061 1211
-89 2977 39
-90 1668 658
-91 878 715
-92 678 1599
-93 1086 868
-94 640 110
-95 3551 1673
-96 106 1267
-97 2243 1332
-98 3796 1401
-99 2643 1320
-100 48 267
