--- sa/input/rat99.tsp.txt
@@ -1,99 +1,0 @@
-  1  6  4
-  2 15 15
-  3 24 18
-  4 33 12
-  5 48 12
-  6 57 14
-  7 67 10
-  8 77 10
-  9 86 15
- 10  6 21
- 11 17 26
- 12 23 25
- 13 32 35
- 14 43 23
- 15 55 35
- 16 65 36
- 17 78 39
- 18 87 35
- 19  3 53
- 20 12 44
- 21 28 53
- 22 33 49
- 23 47 46
- 24 55 52
- 25 64 50
- 26 71 57
- 27 87 57
- 28  4 72
- 29 15 78
- 30 22 70
- 31 34 71
- 32 42 79
- 33 54 77
- 34 66 79
- 35 78 67
- 36 87 73
- 37  7 81
- 38 17 95
- 39 26 98
- 40 32 97
- 41 43 88
- 42 57 89
- 43 64 85
- 44 78 83
- 45 83 98
- 46  5 109
- 47 13 111
- 48 25 102
- 49 38 119
- 50 46 107
- 51 58 110
- 52 67 110
- 53 74 113
- 54 88 110
- 55  2 124
- 56 17 134
- 57 23 129
- 58 36 131
- 59 42 137
- 60 53 123
- 61 63 135
- 62 72 134
- 63 87 129
- 64  2 146
- 65 16 147
- 66 25 153
- 67 38 155
- 68 42 158
- 69 57 154
- 70 66 151
- 71 73 151
- 72 86 149
- 73  5 177
- 74 13 162
- 75 25 169
- 76 35 177
- 77 46 172
- 78 54 166
- 79 65 174
- 80 73 161
- 81 86 162
- 82  2 195
- 83 14 196
- 84 28 189
- 85 38 187
- 86 46 195
- 87 57 194
- 88 63 188
- 89 77 193
- 90 85 194
- 91  8 211
- 92 12 217
- 93 22 210
- 94 34 216
- 95 47 203
- 96 58 213
- 97 66 206
- 98 78 210
- 99 85 204
